extract sort component
